requirements.txt
streamlit
langchain
langchain_openai
openai
python-dotenv
